backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.2
python-jose
python-multipart
python-dotenv
pydantic[email]